get single spacing
