+++ ShapeCrawler/Enums/LineSpacing.cs
@@ -1,0 +1,32 @@
+﻿namespace ShapeCrawler;
+
+/// <summary>
+///     Line Spacing.
+/// </summary>
+public enum LineSpacing
+{
+    /// <summary>
+    ///     Single.
+    /// </summary>
+    Single = 1,
+    
+    /// <summary>
+    ///     1,5 lines 
+    /// </summary>
+    OneHalf = 2,
+    
+    /// <summary>
+    ///     Double.
+    /// </summary>
+    Double = 3,
+    
+    /// <summary>
+    ///     Exactly.
+    /// </summary>
+    Exactly = 4,
+    
+    /// <summary>
+    ///     Multiple.
+    /// </summary>
+    Multiple = 5
+}
